requirements.txt
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
ultralytics
colorthief
pillow
pandas
numpy
scikit-learn
keybert
sentence-transformers
sqlite-utils
streamlit
opencv-python
faiss-cpu
selenium
